One day the constructor will work when I start eclipse

+++ kyoto/src/main/java/uk/ac/ic/kyoto/roguestates/CanadaAgent.java
@@ -15,7 +15,7 @@ import uk.ac.imperial.presage2.util.fsm.FSMException;
 public class CanadaAgent extends NonParticipant {
 
 	public CanadaAgent(UUID id, String name,String ISO, double landArea, double arableLandArea, double GDP,
-			double GDPRate, float availableToSpend, long emissionsTarget, long carbonOffset,
+			double GDPRate, long availableToSpend, long emissionsTarget, long carbonOffset,
 			long energyOutput, long carbonOutput) {
 		super(id, name, ISO, landArea, arableLandArea, GDP,
 				GDPRate, availableToSpend, emissionsTarget, carbonOffset,
